Provide a Git-based implementation of the logics to rebuild the package repository on OpenACS.org
apm_git_build_repository will rebuild either from scratch (clone) or via update (pull) the core repository and all of the non-core packages currently hosted on the GitHub mirror.
A different "channel" is created for every release branch detected in the core repository and for the main branch.

Possible improvements:
 - automated detection of the non-core repos (may require scraping or api integration with the Git mirror)
 - allow to federate a package repository, e.g. allow any OpenACS instance to act as package repository for others

The new proc is currently not used anywhere. The idea is to experiment with it on openacs.org and eventually use it to replace the CVS implementation.