db_multirow: fall back to -local behavior, when used outside of an ADP file
The old behavior was, that when "db_multirow" was called outside an ADP
environment, an error was generated, since the uplevel was determined
without the "-local" flag by [template::adp_level]. Outside ADP,
the result of the function is empty, leading to an error that
a level "#" is invalid. Depending on an optional flag is strange
for a "db_*" command.

Now, the default behavior outside ADP file is local (i.e., when
calling outside ADP, the output variables are set in the calling
scope, unless a different "upvar_level" is specified).