Added support for automatic disconnect when a dynamic cluster node is shut down
When the dynamic cluster configuration has scale for certain
occasions, it makes sense to provide down scaling support when these
occasions are over, which does not rely on the configured cluster
disconnect timeout (ClusterAutodeleteInterval). The new code will send
automatically a disconnect request when a dynamic cluster node is
terminating gracefully.

Depending on the configuration, a new version of NaviServer will be
necessary to reliably execute disconnect requests. Appropriate changes
are in the NaviServer release/4.99 and main branches.

- Bumped version numbers:
  * acs-tcl    to 5.10.1b4
  * acs-admin  to 5.10.1b4