Made "ad_find_all_files" more robust when some subdirectory has too restrictive permissions
Previously, on some OS, Tcl's "glob" command was returning file names,
also, when the permissions should not allow it (macOS), while one
other systems, it raised an error (Linux). The problem is that
"ad_find_all_files" raises an error during startup, not all files are
found, and e.g. some of the files (like e.g. .xql) are not picked up,
... and as a consequence, strange things happen.

With this change, loading is more robust. Now, warnings are issued instead of exceptions.