fix regression test to cope with time zone
Since OpenACS timezone data is now newer as in many tcl versions, some time zones
might not be available in tcl. Therefore, tcl-based conversions with these
time zones will fail.

This change skips these timezones (on my system, 2 cases). But this
is just a part of the fix. The "DB-based conversion" in "lc_time_tz_convert"
seems to be broken for these cases as well, otherwise the regression
test would not have stopped working.