security::validated_host_header: Made acceptance of configured vhosts the first check
Under certain conditions (such as running in a container, or reverse
proxy situations) the admin of a server wants to specify accepted host
names. This can be achieved in the "*/servers" section of a network
driver.  These values are used now first for accepting host header
fields. This change avoids unexpected redirects to, e.g., internal
server addresses.