Made .xql file more consistent by using dot notation
OpenACS for PostgreSQL uses since the release of 5.10.0 the dot
notation for the SQL function acs_permission.permission_p() to ease
portability with Oracle.

In general, one has to be careful that during an upgrade from an older
OpenACS version (e.g. 5.9.*) directly to 5.10.1 to upgrade process
does not depend on the dot notation, otherwise the upgrade will fail.
One should be safe for most UI functions in this respect.