Fixed issue introduced in OpenACS 5.9.0
The old version did not insert a property value via the
sec_session_property__upsert() in PostgreSQL on the initial setting
(later updates were OK). The broken version was just adding a tuple
and left the "property_value" empty.

Many thanks to Jonathan Kelley for identifying the issue and reporting it.