::acs_mail_lite::configured_p: check, if we can connect to the mail server
This new function has the goal to centralize tests to check, whether
the SMTP server is properly configured and ready to accept mail
requests. This is currently just used in the regression test to skip
hopeless tests.

Later, we should extend the regression test to test whether
sending of real emails works.