fixed test server_startup_ok
Previously, the test was returning different results depending on the order of
tests.

While it is in general better to avoid error entries in the log file during
regression test, since these testing expected behavior, while error entries
in the log file should indicated unexpected behavior. However, it is still
a long way to let the regression test run cleanly.