New pages for admins: Security and Privacy Posture Overview
As expressed as a wish from OpenACS users at the last OpenACS
conference, a "Security and Privacy Posture Overview" was added that
offers a quick overview of the state of the system and eases access to
the parameters scattered over different packages in the system.

The page offers:
- Quick overview
- Check of security and privacy relevant package parameters
- Permission and accessibility check of mounted packages
- Response header check
- External library check (CDN vs local usage, vulnerable or outdated libraries)

TODO: One should probably reconsider the permissions of some of the standard site nodes
(similar to what we did with the API browser some time ago).