New API call: aa_silence_log_entries
The new function silences expected messages in the system log. The
proc deactivates the specified severity levels during the code in the
last arguemnt is executed. After it has finished, the severity levels
are reset to their previous values.

The purpose of this function is to make test system log clean from
error messages and the like during the regression test, such it
contains only errors which have to be looked at.