Fixed incorrect result handling from ::acs::site_node get_node_id
When

    ::acs::site_node get_node_id -url /$mount_point

is executed with a non-existing mount point, the node info of the
parent_id is returned. The code assumed that for such cases, the
resulting "object_id" member is empty. Since this affects only the
two action procs

     install::xml::action::mount
     install::xml::action::mount-existing

i have adjusted the behavior there.
Furthermore, some tests are now less mind-twisted.